requirements.txt
discord.py
python-dotenv
python-steam-api
requests